+++ src/Core/GlobalData.m
@@ -102,7 +102,7 @@ static GlobalData *_privateDataSingleton = nil;
 }
 
 
--(unsigned long)retainCount {
+-(NSUInteger)retainCount {
     return UINT_MAX;  //denotes an object that cannot be release
 }
